+++ src/components/LicenseForm/LicenseForm.js
@@ -10,8 +10,11 @@ import {
 
 import {
   LicenseFormInfo,
+  LicenseFormCoreDocs,
   LicenseFormTerms,
-} from './Sections';
+} from './sections';
+
+import css from './LicenseForm.css';
 
 class LicenseForm extends React.Component {
   static propTypes = {
@@ -22,7 +25,8 @@ class LicenseForm extends React.Component {
   state = {
     sections: {
       licenseFormInfo: true,
-      licenseFormTerms: true,
+      licenseFormDocs: false,
+      licenseFormTerms: false,
     }
   }
 
@@ -52,18 +56,21 @@ class LicenseForm extends React.Component {
     const { sections } = this.state;
 
     return (
-      <AccordionSet>
-        <Row end="xs">
-          <Col xs>
-            <ExpandAllButton
-              accordionStatus={sections}
-              onToggle={this.handleAllSectionsToggle}
-            />
-          </Col>
-        </Row>
-        <LicenseFormInfo id="licenseFormInfo" open={sections.licenseFormInfo} {...sectionProps} />
-        <LicenseFormTerms id="licenseFormTerms" open={sections.licenseFormTerms} {...sectionProps} />
-      </AccordionSet>
+      <div className={css.licenseForm}>
+        <AccordionSet>
+          <Row end="xs">
+            <Col xs>
+              <ExpandAllButton
+                accordionStatus={sections}
+                onToggle={this.handleAllSectionsToggle}
+              />
+            </Col>
+          </Row>
+          <LicenseFormInfo id="licenseFormInfo" open={sections.licenseFormInfo} {...sectionProps} />
+          <LicenseFormCoreDocs id="licenseFormDocs" open={sections.licenseFormDocs} {...sectionProps} />
+          <LicenseFormTerms id="licenseFormTerms" open={sections.licenseFormTerms} {...sectionProps} />
+        </AccordionSet>
+      </div>
     );
   }
 }