+++ src/components/LicenseForm/components/index.js
@@ -1,0 +1,2 @@
+export { default as OrganizationsFieldArray } from './OrganizationsFieldArray';
+export { default as TermsListField } from './TermsListField';