+++ src/components/ViewLicense/sections/LicenseOrganizations.js
@@ -5,7 +5,7 @@ import { FormattedMessage } from 'react-intl';
 import { Link } from 'react-router-dom';
 import { Accordion, Badge, Icon, Layout } from '@folio/stripes/components';
 
-export default class Organizations extends React.Component {
+export default class LicenseOrganizations extends React.Component {
   static propTypes = {
     license: PropTypes.shape({
       orgs: PropTypes.arrayOf(
@@ -59,6 +59,7 @@ export default class Organizations extends React.Component {
   render() {
     return (
       <Accordion
+        closedByDefault
         displayWhenClosed={this.renderBadge()}
         displayWhenOpen={this.renderBadge()}
         id="license-orgs"