+++ src/index.js
@@ -34,3 +34,4 @@ class App extends React.Component {
 }
 
 export default App;
+export { Licenses };